adding main stuff
