+++ src/main/java/com/mysite/pfarmscripts/resources/User.java
@@ -1,0 +1,11 @@
+package com.mysite.pfarmscripts.resources;
+
+
+
+/**
+ * Created by Val on 8/4/2016.
+ */
+public class User {
+    public static String UserName = "MMMuse";
+    public static String Password = "saber12994";
+}
