--- src/test/java/com/mysite/pfarmscripts/User.java
@@ -1,9 +1,0 @@
-package com.mysite.pfarmscripts;
-
-/**
- * Created by Val on 8/4/2016.
- */
-public class User {
-    public static String UserName = "valaadus";
-    public static String Password = "bakednickthankfuldrag";
-}
